+++ CHANGELOG.md
@@ -5,6 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.7.0 - 2021-10-19
+
+# Added
+
+  * Scripts system and `flask script`.
+  * `gen_dhcp.py` script for generating DHCP hosts rules, called at
+    device register / transfer.
+
+# Changed
+
+  * Rezidents devices are sorted by ID.
+
+# Fixed
+
+  * Profile page crashed when several devices.
+
+
 ## 0.6.1 - 2021-10-18
 
 # Changed